EFS Ransomware workaround
Research and original story:
https://safebreach.com/Post/EFS-Ransomware

Background:
https://www.zdnet.com/article/antivirus-vendors-scramble-to-fix-new-efs-ransomware-attack/

Microsoft and Avira are NOT planning to fix it since it's not a "Security Servicing Criteria" (according to MS). Other AV's are listed under "Vendor Status" and most AV's trying to patch/detect this. Windows own Defender (CFA) is not going to detect it, what Microsoft will do instead is to detect the malware via signature updates in their reputation database.

The EFS toggle is explained officially here:
https://docs.microsoft.com/en-us/openspecs/windows_protocols/ms-gpef/0382ec4d-bfa9-46c9-a99a-1f2e042938c0

In case you use/need EFS you can enable Data Recovery Agent (DRA) to workaround it, it's explained here:
https://docs.microsoft.com/en-us/windows/security/information-protection/windows-information-protection/create-and-verify-an-efs-dra-certificate

I do not provide an undo (default) registry file because I believe my followers should get the maximum security and migrate this ASAP.

EFS is known to be "problematic":
https://en.wikipedia.org/wiki/Encrypting_File_System
https://medium.com/m/global-identity?redirectUrl=https%3A%2F%2Fdcnoren.com%2Fattacking-windows-efs-with-metasploit-69a7c92e0d57
https://old.reddit.com/r/crypto/comments/8qqw4w/is_microsofts_encrypting_file_system_efs_secure/?st=k5qphk9x&sh=46a259c1