(Unofficial) Microsoft KB Search project mentioned
Announce:
https://www.pkisolutions.com/mskb-archive-launch/

Stats (as of today):
- Total articles from Microsoft – 49,430
- Active articles on Microsoft.com – 48,646
- Archived articles in our MSKB – 784

Microsoft often removes or changes the URL of KB's (with "often" I mean here and then out of the blue). A user finally provided a database to archive the entries for reference reasons. Another way would be to use web.archive.org, the problem here is also that the website is not perfect nor reliable in terms of finding specific KB's.

I guess the project is useful even if it's not official from MS and that's the reason why it belongs on the front page.  Usability.