Enable/Disable Get Even More Out of Windows
This tweak is only for reference, it was removed in 1709/1803 or 1809 (since I forgot which Build exactly I put it under the "All Windows versions" category (it also does not "hurt" if you accidentally applied it).

If someone knows when it was removed from Windows, let me know and I move it in the correct section.