Enable/Disable external link handler in Chrome
## Overview
Due to security reasons this was disabled and you manually need to approve external request from applications/websites.

## Details
When an external website or application wants to access specific flagged protocols you will see a popup which tells you that xyz website/app want to open it. You need to manually approve it, which can be annoying but it's done to block malicous requests.